pyproject.toml
[project]
name = "tfg-ui-dfrf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dashscope>=1.25.15",
    "dotenv>=0.9.9",
    "flask>=3.1.3",
    "pygithub>=2.9.0",
]
